+++ portal-impl/src/com/liferay/portlet/bookmarks/service/impl/BookmarksFolderLocalServiceImpl.java
@@ -14,11 +14,13 @@
 
 package com.liferay.portlet.bookmarks.service.impl;
 
+import com.liferay.portal.kernel.dao.orm.QueryDefinition;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.model.ResourceConstants;
 import com.liferay.portal.model.User;
 import com.liferay.portal.service.ServiceContext;
@@ -27,6 +29,7 @@ import com.liferay.portlet.bookmarks.model.BookmarksEntry;
 import com.liferay.portlet.bookmarks.model.BookmarksFolder;
 import com.liferay.portlet.bookmarks.model.BookmarksFolderConstants;
 import com.liferay.portlet.bookmarks.service.base.BookmarksFolderLocalServiceBaseImpl;
+import com.liferay.portlet.trash.model.TrashEntry;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -81,13 +84,33 @@ public class BookmarksFolderLocalServiceImpl
 	public void deleteFolder(BookmarksFolder folder)
 		throws PortalException, SystemException {
 
+		deleteFolder(folder, true);
+	}
+
+	public void deleteFolder(
+			long folderId, boolean includeTrashedEntries)
+		throws PortalException, SystemException {
+
+		BookmarksFolder folder = bookmarksFolderLocalService.getFolder(
+			folderId);
+
+		deleteFolder(folder, includeTrashedEntries);
+	}
+
+	public void deleteFolder(
+			BookmarksFolder folder, boolean includeTrashedEntries)
+		throws PortalException, SystemException {
+
 		// Folders
 
-		List<BookmarksFolder> folders = bookmarksFolderPersistence.findByG_P(
-			folder.getGroupId(), folder.getFolderId());
+		List<BookmarksFolder> folders = bookmarksFolderPersistence.findByG_P_S(
+			folder.getGroupId(), folder.getFolderId(),
+			WorkflowConstants.STATUS_ANY);
 
 		for (BookmarksFolder curFolder : folders) {
-			deleteFolder(curFolder);
+			if (includeTrashedEntries || !curFolder.isInTrash()) {
+				deleteFolder(curFolder);
+			}
 		}
 
 		// Folder
@@ -102,7 +125,7 @@ public class BookmarksFolderLocalServiceImpl
 		// Entries
 
 		bookmarksEntryLocalService.deleteEntries(
-			folder.getGroupId(), folder.getFolderId());
+			folder.getGroupId(), folder.getFolderId(), includeTrashedEntries);
 
 		// Expando
 
@@ -114,6 +137,11 @@ public class BookmarksFolderLocalServiceImpl
 		subscriptionLocalService.deleteSubscriptions(
 			folder.getCompanyId(), BookmarksFolder.class.getName(),
 			folder.getFolderId());
+
+		// Trash
+
+		trashEntryLocalService.deleteEntry(
+			BookmarksFolder.class.getName(), folder.getFolderId());
 	}
 
 	public void deleteFolder(long folderId)
@@ -170,14 +198,69 @@ public class BookmarksFolderLocalServiceImpl
 			long groupId, long parentFolderId, int start, int end)
 		throws SystemException {
 
-		return bookmarksFolderPersistence.findByG_P(
-			groupId, parentFolderId, start, end);
+		return getFolders(
+			groupId, parentFolderId, WorkflowConstants.STATUS_APPROVED, start,
+			end);
+	}
+
+	public List<BookmarksFolder> getFolders(
+			long groupId, long parentFolderId, int status, int start, int end)
+		throws SystemException {
+
+		return bookmarksFolderPersistence.findByG_P_S(
+				groupId, parentFolderId, status, start, end);
+	}
+
+	public List<Object> getFoldersAndEntries(long groupId, long folderId)
+		throws SystemException {
+
+		return getFoldersAndEntries(
+			groupId, folderId, WorkflowConstants.STATUS_ANY);
+	}
+
+	public List<Object> getFoldersAndEntries(
+			long groupId, long folderId, int status)
+		throws SystemException {
+
+		QueryDefinition queryDefinition = new QueryDefinition(status);
+
+		return bookmarksFolderFinder.findBF_BE_ByG_F(
+			groupId, folderId, queryDefinition);
+	}
+
+	public List<Object> getFoldersAndEntries(
+			long groupId, long folderId, int status, int start, int end)
+		throws SystemException {
+
+		QueryDefinition queryDefinition = new QueryDefinition(
+			status, start, end, null);
+
+		return bookmarksFolderFinder.findBF_BE_ByG_F(
+			groupId, folderId, queryDefinition);
+	}
+
+	public int getFoldersAndEntriesCount(
+			long groupId, long folderId, int status)
+		throws SystemException {
+
+		QueryDefinition queryDefinition = new QueryDefinition(status);
+
+		return bookmarksFolderFinder.countBF_BE_ByG_F(
+				groupId, folderId, queryDefinition);
 	}
 
 	public int getFoldersCount(long groupId, long parentFolderId)
 		throws SystemException {
 
-		return bookmarksFolderPersistence.countByG_P(groupId, parentFolderId);
+		return getFoldersCount(
+			groupId, parentFolderId, WorkflowConstants.STATUS_APPROVED);
+	}
+
+	public int getFoldersCount(long groupId, long parentFolderId, int status)
+		throws SystemException {
+
+		return bookmarksFolderPersistence.countByG_P_S(
+				groupId, parentFolderId, status);
 	}
 
 	public void getSubfolderIds(
@@ -193,6 +276,49 @@ public class BookmarksFolderLocalServiceImpl
 			getSubfolderIds(
 				folderIds, folder.getGroupId(), folder.getFolderId());
 		}
+	}
+
+	public BookmarksFolder moveFolder(long folderId, long parentFolderId)
+		throws PortalException, SystemException {
+
+		BookmarksFolder folder = bookmarksFolderPersistence.findByPrimaryKey(
+			folderId);
+
+		folder.setParentFolderId(parentFolderId);
+
+		bookmarksFolderPersistence.update(folder);
+
+		return folder;
+	}
+
+	public BookmarksFolder moveFolderFromTrash(
+			long userId, long folderId, long parentFolderId)
+		throws PortalException, SystemException {
+
+		restoreFolderFromTrash(userId, folderId);
+
+		return moveFolder(folderId, parentFolderId);
+	}
+
+	public void moveFolderToTrash(long userId, long folderId)
+		throws PortalException, SystemException {
+
+		BookmarksFolder folder = bookmarksFolderPersistence.findByPrimaryKey(
+			folderId);
+
+		updateStatus(userId, folder, WorkflowConstants.STATUS_IN_TRASH);
+	}
+
+	public void restoreFolderFromTrash(long userId, long folderId)
+		throws PortalException, SystemException {
+
+		BookmarksFolder folder = bookmarksFolderPersistence.findByPrimaryKey(
+			folderId);
+
+		TrashEntry trashEntry = trashEntryLocalService.getEntry(
+			BookmarksFolder.class.getName(), folderId);
+
+		updateStatus(userId, folder, trashEntry.getStatus());
 	}
 
 	public void subscribeFolder(long userId, long groupId, long folderId)
@@ -246,6 +372,46 @@ public class BookmarksFolderLocalServiceImpl
 		folder.setExpandoBridgeAttributes(serviceContext);
 
 		bookmarksFolderPersistence.update(folder);
+
+		return folder;
+	}
+
+	public BookmarksFolder updateStatus(
+			long userId, BookmarksFolder folder, int status)
+		throws PortalException, SystemException {
+
+		int oldStatus = folder.getStatus();
+
+		User user = userPersistence.findByPrimaryKey(userId);
+
+		Date now = new Date();
+
+		folder.setStatus(status);
+		folder.setStatusByUserId(userId);
+		folder.setStatusByUserName(user.getFullName());
+		folder.setStatusDate(now);
+
+		bookmarksFolderPersistence.update(folder);
+
+		// BookmarksFolders and BookmarksEntries
+
+		List<Object> foldersAndEntries =
+			bookmarksFolderLocalService.getFoldersAndEntries(
+				folder.getGroupId(), folder.getFolderId());
+
+		updateDependentStatus(foldersAndEntries, status);
+
+		// Trash
+
+		if (oldStatus == WorkflowConstants.STATUS_IN_TRASH) {
+			trashEntryLocalService.deleteEntry(
+				BookmarksFolder.class.getName(), folder.getFolderId());
+		}
+		else if (status == WorkflowConstants.STATUS_IN_TRASH) {
+			trashEntryLocalService.addTrashEntry(
+				userId, folder.getGroupId(), BookmarksFolder.class.getName(),
+				folder.getFolderId(), oldStatus, null, null);
+		}
 
 		return folder;
 	}
@@ -331,6 +497,88 @@ public class BookmarksFolderLocalServiceImpl
 		}
 
 		deleteFolder(fromFolder);
+	}
+
+	protected void updateDependentStatus(
+			List<Object> foldersAndEntries, int status)
+		throws PortalException, SystemException {
+
+		for (Object object : foldersAndEntries) {
+			if (object instanceof BookmarksEntry) {
+				BookmarksEntry entry = (BookmarksEntry)object;
+
+				if (status == WorkflowConstants.STATUS_IN_TRASH) {
+					if (entry.getStatus() ==
+							WorkflowConstants.STATUS_APPROVED) {
+
+						assetEntryLocalService.updateVisible(
+							BookmarksEntry.class.getName(), entry.getEntryId(),
+							false);
+					}
+
+					// Social
+
+					socialActivityCounterLocalService.disableActivityCounters(
+						BookmarksEntry.class.getName(), entry.getEntryId());
+
+					// Index
+
+					Indexer indexer = IndexerRegistryUtil.nullSafeGetIndexer(
+						BookmarksEntry.class);
+
+					indexer.reindex(entry);
+
+					if (entry.getStatus() == WorkflowConstants.STATUS_PENDING) {
+						entry.setStatus(WorkflowConstants.STATUS_DRAFT);
+
+						bookmarksEntryPersistence.update(entry);
+
+						workflowInstanceLinkLocalService.
+							deleteWorkflowInstanceLink(
+								entry.getCompanyId(), entry.getGroupId(),
+								BookmarksEntry.class.getName(),
+								entry.getEntryId());
+					}
+				}
+				else {
+
+					// Asset
+
+					if (entry.getStatus() ==
+							WorkflowConstants.STATUS_APPROVED) {
+
+						assetEntryLocalService.updateVisible(
+							BookmarksEntry.class.getName(), entry.getEntryId(),
+							true);
+					}
+
+					// Social
+
+					socialActivityCounterLocalService.enableActivityCounters(
+						BookmarksEntry.class.getName(), entry.getEntryId());
+
+					// Index
+
+					Indexer indexer = IndexerRegistryUtil.nullSafeGetIndexer(
+						BookmarksEntry.class);
+
+					indexer.reindex(entry);
+				}
+			}
+			else if (object instanceof BookmarksFolder) {
+				BookmarksFolder folder = (BookmarksFolder)object;
+
+				if (folder.isInTrash()) {
+					continue;
+				}
+
+				List<Object> curFoldersAndEntries =
+					bookmarksFolderLocalService.getFoldersAndEntries(
+						folder.getGroupId(), folder.getFolderId());
+
+				updateDependentStatus(curFoldersAndEntries, status);
+			}
+		}
 	}
 
 	protected void validate(String name) throws PortalException {
